Move header-altering helper to module generic helpers
This helper is for editing the header tags in some module-independent
way, so this piece of code is now moved to the module-generic helpers.
